Rename module jpro-scroll -> jpro-sticky
jpro-scroll collided with jpro-html-scrollpane. jpro-sticky names the
headline feature (position: sticky) and is what users search for; fixed
positioning rides along as sticky's sibling, the way jpro-file covers
open and save.

Renames the directory, package (one.jpro.platform.scroll ->
one.jpro.platform.sticky), JPMS module, publish wiring, and the module's
log/DOM identity tokens. The public Scroll/ScrollPosition/ScrollAnchor
class names are kept as-is (neutral umbrella over sticky + fixed).

Also finalizes README.md to document the ScrollAnchor per-axis model.
