--- jpro-scroll/example/build.gradle
@@ -1,21 +1,0 @@
-plugins {
-    id 'org.openjfx.javafxplugin'
-    id 'jpro-gradle-plugin'
-}
-
-dependencies {
-    implementation(project(":jpro-scroll"))
-    implementation "io.github.mkpaz:atlantafx-base:$ATLANTAFX_VERSION"
-    implementation "org.slf4j:slf4j-api:$SLF4J_API_VERSION"
-    runtimeOnly "ch.qos.logback:logback-classic:$LOGBACK_VERSION"
-}
-
-javafx {
-    version = "$JAVAFX_EXAMPLES_VERSION"
-    modules = ['javafx.controls']
-}
-
-application {
-    mainClass = 'one.jpro.platform.scroll.example.ScrollSample'
-    mainModule = "one.jpro.platform.scroll.example"
-}
