--- jpro-scroll/example/src/main/java/module-info.java
@@ -1,11 +1,0 @@
-/**
- * Module descriptor for the example application.
- *
- * @author Tobias Horak
- */
-module one.jpro.platform.scroll.example {
-    requires javafx.controls;
-    requires one.jpro.platform.scroll;
-
-    exports one.jpro.platform.scroll.example;
-}
