+++ jpro-sticky/example/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+/**
+ * Module descriptor for the example application.
+ *
+ * @author Tobias Horak
+ */
+module one.jpro.platform.sticky.example {
+    requires javafx.controls;
+    requires one.jpro.platform.sticky;
+
+    exports one.jpro.platform.sticky.example;
+}
