+++ settings.gradle
@@ -20,8 +20,8 @@ include "jpro-session"
 include "jpro-session:example"
 include "jpro-html-scrollpane"
 include "jpro-html-scrollpane:example"
-include "jpro-scroll"
-include "jpro-scroll:example"
+include "jpro-sticky"
+include "jpro-sticky:example"
 include "jpro-youtube"
 include "jpro-youtube:example"
 include "jpro-webrtc"
